move client to config.h
